Locked in ceks

This function takes a Raster object and identifies non NA cells and writes them to a .dat file as a Must Include. The file will be written to the location specified by the name argument. If the file already exists, it will be overwritten. The file format is plain text, with each line terminated by a newline character.

Usage

Locked_in_Cells(Lockedin, name = "Problem")

Arguments

Lockedin

A Raster object with a value of one in the cells that are locked in as solutions to the problem and NA values elsewhere

name

The name of the output file

Value

.dat file. This function is used for the side-effect of writing values to a file.

Author

Derek Corcoran

Examples

data(Species)
library(terra)
#> terra 1.8.54
Test <- Species[[1]] |>
terra::unwrap()

# Generate the "Problem.dat" file

define_cells(Test[[1]])
#>           used (Mb) gc trigger  (Mb) max used  (Mb)
#> Ncells 1748552 93.4    3301388 176.4  3214774 171.7
#> Vcells 2625640 20.1    8388608  64.0  5371240  41.0

file.remove("Problem.dat")
#> [1] TRUE