Landuse names

This function takes a vector of landuse names and writes them to a .dat file. The file will be written to the location specified by the name argument. If the file already exists, it will be overwritten. The file format is plain text, with each line terminated by a newline character.

Usage

landuse_names(landuses = NULL, name = "Problem")

Arguments

landuses

a vector with the names of the landuses

name

The name of the output file

Value

.dat file. This function is used for the side-effect of writing values to a file.

Author

Derek Corcoran

Examples

landuse_names(landuses =  c("Agriculture", "Forest", "Urban"))
#>           used (Mb) gc trigger  (Mb) max used  (Mb)
#> Ncells 1782863 95.3    3301388 176.4  3301388 176.4
#> Vcells 3180818 24.3   10146329  77.5 10117639  77.2

# delete the file so the test on cran can pass this

file.remove("Problem.dat")
#> [1] TRUE